requirements.txt
PyYAML
aiohttp
loguru
pydantic 